Document 2026-09-08 findings, analysis decisions, and rig settings
Notes from the first full acquisition day, and the conventions the analysis
package now follows.

docs/session_notes_20260908.md: what was changed each session and what
resulted. The DAC commands were correct throughout; every failure was
downstream. Two rig faults account for all the apparent crosstalk -- the 565
driver left at minimum power from 19:58, and 565 detector gain high enough to
amplify ordinary 24% optical bleed into an apparent 300-900x contamination.
Records the dF/F comparison across excitation wavelengths, the three envelope
oscillations and where each enters, and the outstanding rig work.

docs/analysis_decisions.md: why carrier QC comes first, why the demodulation
window needs 8 bins of carrier separation rather than 3, why the envelope is
low-passed rather than notched, how consumption licks are bounded by the next
cue, and when to use dF/F versus z-score. Includes the pitfalls found the hard
way: rail detection anchored to a transient maximum, filtfilt padding far too
short for narrow filters, artefact prominence measured in a band the filter had
already removed, and high SNR on a dark channel.

docs/empirical_carrier_qc.md: corrects the PS111 bright-470 conclusion. The dB
figures replicate, but the carriers were 35 mV and 1.9 mV on channels sitting
at -0.207 V and -0.335 V, i.e. dark. No PS111 or PS112 recording contains a
usable carrier.

README: document the analysis package and entry points; add bench notes on
carrier amplitude, choosing carriers away from 60 Hz harmonics, and checking
detector gain.

Also applies ruff fixes across the analysis package.

+++ src/labjack_photometry_gui/analysis/align.py
@@ -82,8 +82,8 @@ def align_to_events(
         return AlignedTraces(empty, np.zeros(0), np.zeros(0), np.zeros(0, dtype=int))
 
     rate = 1.0 / float(np.median(np.diff(time_s)))
-    before = int(round(pre_s * rate))
-    after = int(round(post_s * rate))
+    before = round(pre_s * rate)
+    after = round(post_s * rate)
     offsets = np.arange(-before, after + 1)
     axis = offsets / rate
